+++ src/execute.js
@@ -1,77 +1,32 @@
 const vm = require('vm');
-const path = require('path');
-
-/**
- * 
- * @param {string} parseStatement 
- *  string of require statement (e.g. const a = require('module'))
- * @param {object} sandbox variable environtment to execute upon
- * @param {string} basePath base path to which the require will be relative
- * @return {object} updated sandbox
- */
-function executeRequireStatement(parseStatement, sandbox, basePath) {
-  const requireParseRegex = /require\(['"](.*?)['"]\)(.*)/;
-
-  const requireInfo = requireParseRegex.exec(parseStatement);
-  const pathToRequire = requireInfo[1]; // string inside require('')
-
-  // String after require to handle cases like require('module').foo
-  const codeAfterRequire = requireInfo[2] || '';
-
-  let temp;
-
-  if (pathToRequire.startsWith('.')) {
-    // path is a local file
-    temp = require(
-      path.join(
-        basePath,
-        pathToRequire
-      )
-    );
-  } else {
-    // path is a nodejs module
-    temp = require(pathToRequire);
-  }
-
-
-  const context = {temp};
-  vm.createContext(context);
-
-  vm.runInContext(
-    parseStatement
-      .slice(0, parseStatement.indexOf('='))
-      .replace(/(?:const |let )/g, 'var ')
-      .trim() + 
-    ' = temp' + codeAfterRequire,
-    context
-  );
-
-  delete context.temp; // delete the temporary created variable
-
-  return {...sandbox, ...context};
-}
-
 
 /**
  * Executes the JavaScript code from string
- * 
- * @param {string} jsToExecute 
+ *
+ * @param {string} jsToExecute
  *  JavaScript code in String to parse and execute
- * @param {any} sandbox  
+ * @param {any} sandbox
  *  The variables and all the information required by program
  * @return {object} {type, sandbox, value?}
  */
 function execute(jsToExecute, sandbox = {}) {
   const numOfKeysBeforeExecution = Object.keys(sandbox).length;
-  const codeToExecute = 
-    'aBellSpecificVariable = ' + 
+  const codeToExecute =
+    'aBellSpecificVariable = ' +
     jsToExecute.replace(/(?:const |let |var )/g, '');
-    
+
   const script = new vm.Script(codeToExecute);
   const context = new vm.createContext(sandbox); // eslint-disable-line
-  script.runInContext(context);
-  
-  const sandboxVariable = sandbox.aBellSpecificVariable; 
+
+  try {
+    script.runInContext(context, {
+      displayErrors: true
+    });
+  } catch (err) {
+    throw new Error(err);
+  }
+
+  const sandboxVariable = sandbox.aBellSpecificVariable;
 
 
   // Add Console Log Logic
@@ -85,10 +40,11 @@ function execute(jsToExecute, sandbox = {}) {
     // A chance of script being an assignment
 
     /**
-     *  if new variable is added to the context, 
+     *  if new variable is added to the context,
      *  (which means the script added a new variable and was an assignment)
-    */ 
-    if ((Object.keys(sandbox).length - 1) > numOfKeysBeforeExecution) {
+     */
+
+    if (Object.keys(sandbox).length - 1 > numOfKeysBeforeExecution) {
       delete sandbox.aBellSpecificVariable;
       return {
         type: 'assignment',
@@ -97,15 +53,21 @@ function execute(jsToExecute, sandbox = {}) {
     }
 
     // if a predefined variable is assigned a new value, it will escape the conditions above
-    const textToLookEqualSignIn = jsToExecute.replace(/`(?:.*?)`|"(?:.*?)"|'(?:.*?)'/gs, '');
+
+    const textToLookEqualSignIn = jsToExecute.replace(
+      /`(?:.*?)`|"(?:.*?)"|'(?:.*?)'/gs,
+      ''
+    ); // removes strings to avoid checking equal-to signs inside string
+
+    // Look for word-equalsign-word
     if (textToLookEqualSignIn.match(/[\w\d ]={1}(?![>=])/g) !== null) {
-      // It is 90% chance that it is assignment
+      // 90% chance that this is an assignment
       delete sandbox.aBellSpecificVariable;
       return {
         type: 'assignment',
         sandbox
       };
-    } 
+    }
   }
   // script is not assignment i.e it returns some value that can be printed
   return {
@@ -115,7 +77,4 @@ function execute(jsToExecute, sandbox = {}) {
   };
 }
 
-module.exports = {
-  execute,
-  executeRequireStatement
-};
+module.exports = execute;