+++ CHANGELOG.md
@@ -1,14 +1,53 @@
 # Changelog
 
+## v0.1.12
+
+- Throw error at `execute` to avoid having vm data in error stack
+
+## v0.1.11
+
+- Trim value before adding.
+
+## v0.1.10
+
+- **BREAKING CHANGE**, **SECURITY UPDATE**
+  To use `require()` in the template, user will have to pass `allowRequire: true` in option. This option is by default set to false.
+  ```js
+  const newHTMLTemplate = abellRenderer.render(myAbellTemplate, mySandbox, {
+    allowRequire: true
+  });
+  ```
+
+## v0.1.9
+
+- Fix to recursively find and create nested `.abell` files
+
+## v0.1.8
+
+- basePath is set to paths relative to input files.
+
+## v0.1.7
+
+- **BUG FIX**
+  Allowing CLI to build without output path
+
+## v0.1.6
+
+- **MAJOR CHANGE**
+  Added ability to escape the brackets with a slash('\')
+
 ## v0.1.5
+
 - Support for expression after `require` (e.g. `require('module1').someProperty`)
 - Build a folder with CLI (Issue: [#6](https://github.com/abelljs/abell-renderer/issues/6), PR: [#8](https://github.com/abelljs/abell-renderer/pull/8), Thanks to [@Pika1998](https://github.com/Pika1998))
 
 ## v0.1.4
+
 - Error handling when the temlate does not have any `{{` `}}`
 - More tests written for `render()` function
 
 ## v0.1.3
+
 - Refactoring of major logic
 - Conditions executed as assignment bug fixed
 - More tests added in Unit Testing.
@@ -17,28 +56,35 @@
 - Template build time added to CLI
 
 ## v0.1.2
+
 BREAKING CHANGE
+
 - `abell-renderer [flags]` changed to `abell-renderer build [flags]`
 - `help, --version` added.
 
 ## v0.1.1
+
 Bug that stored last value and printed it when assignment was used, fixed.
 
 ## v0.1.0
+
 Better log messages added in CLI
 
 ## v0.1.0-beta.1
+
 - CLI added to render `.abell` files
 
 ## v0.1.0-alpha.2
+
 - Support for nodejs modules
 
 ## v0.1.0-alpha.1
+
 - Sequential compilation of JavaScript [#2](https://github.com/abelljs/abell-renderer/issues/2)
 - Require values from json/js in .abell files
 
-
 ## v0.0.3
+
 Ability to add functions inside curly brackets
 
 ```js
@@ -50,6 +96,7 @@ Ability to add functions inside curly brackets
 ```
 
 outputs:
+
 ```html
 <div>
   8
@@ -57,4 +104,5 @@ outputs:
 ```
 
 ## < v0.0.2
+
 Initilization and overall setup